refactor: 컴포넌트 파일들을 components 폴더 하위로 이동
