+++ src/index.js
@@ -2,7 +2,9 @@ import '../templates/style.css';
 
 import LunchMenuController from './controller/LunchMenuController';
 
-import RestaurantList from './view/RestaurantList';
-import RestaurantListItem from './view/RestaurantListItem';
+import RestaurantList from './view/components/RestaurantList';
+import RestaurantListItem from './view/components/RestaurantListItem';
+import RestaurantRegisterModal from './view/components/RestaurantRegisterModal';
 
 LunchMenuController.render();
+LunchMenuController.bindEvents();
